fix: create Extension classes
