--- packages/cli-plugin-scaffold-extensions/src/generateExtension/generators/index.ts
@@ -1,8 +1,0 @@
-import { apiGenerator } from "./api";
-import { adminGenerator } from "./admin";
-import { PluginGenerator } from "~/types";
-
-export const generators: Record<string, PluginGenerator> = {
-    api: apiGenerator,
-    admin: adminGenerator
-};
