--- packages/cli-plugin-scaffold-extensions/src/generateExtension/generators/utils/addPluginToApiApp.ts
@@ -1,53 +1,0 @@
-import { Project, ArrayLiteralExpression, Node } from "ts-morph";
-import path from "path";
-import { formatCode } from "@webiny/cli-plugin-scaffold/utils";
-
-interface Params {
-    name: string;
-    packageName: string;
-}
-
-export const addPluginToApiApp = async (params: Params): Promise<void> => {
-    const { name, packageName } = params;
-
-    const extensionsFilePath = path.join("apps", "api", "graphql", "src", "extensions.ts");
-
-    const extensionFactory = name + "ExtensionFactory";
-    const importName = "{ createExtension as " + extensionFactory + " }";
-    const importPath = packageName;
-
-    const project = new Project();
-    project.addSourceFileAtPath(extensionsFilePath);
-
-    const source = project.getSourceFileOrThrow(extensionsFilePath);
-
-    const existingImportDeclaration = source.getImportDeclaration(importPath);
-    if (existingImportDeclaration) {
-        throw new Error(
-            `Could not import  "${importPath}" in "${extensionsFilePath}" as it already exists.`
-        );
-    }
-
-    let index = 1;
-
-    const importDeclarations = source.getImportDeclarations();
-    if (importDeclarations.length) {
-        const last = importDeclarations[importDeclarations.length - 1];
-        index = last.getChildIndex() + 1;
-    }
-
-    source.insertImportDeclaration(index, {
-        defaultImport: importName,
-        moduleSpecifier: importPath
-    });
-
-    const pluginsArray = source.getFirstDescendant(node =>
-        Node.isArrayLiteralExpression(node)
-    ) as ArrayLiteralExpression;
-
-    pluginsArray.addElement(`${extensionFactory}()`);
-
-    await source.save();
-
-    await formatCode(extensionsFilePath, {});
-};
