+++ packages/cli-plugin-scaffold-extensions/src/types.ts
@@ -1,15 +1,3 @@
-interface PluginGeneratorParams {
-    input: {
-        name: string;
-        location: string;
-        packageName: string;
-    };
-}
-
-export type PluginGenerator = (params: PluginGeneratorParams) => Promise<{
-    nextSteps?: string[];
-}>;
-
 export interface Input {
     type: string;
     name: string;
@@ -17,11 +5,4 @@ export interface Input {
     location: string;
     dependencies?: string;
     templateArgs?: string;
-}
-
-export interface DownloadedExtensionData {
-    folderName: string;
-    folderPath: string;
-    packageJsonPath: string;
-    extensionType: string | null;
 }
